feat: use simplex views for registry and update index handling in registry (#29)
There are three places where the view is incremented in threshold simplex:
handle_notarization
handle_finalization
handle_nullification

The happy path is handle_notarization -> handle_finalization.
However, in case of a nullification, there is no block added, but the view is still incremented.

This causes the view to outpace the block height. This isn't an issue in itself, however, we've been using the block height to update registry. But simplex will use the view when it calls the registry methods (via the Supervisor traits).

This can cause the following issue. For example, lets say we are at block height 4 and view 6, with 10 validators:
- A new round is created for view 6.
- It will call the participants(6) method with view 6, which will return a list of the 10 validators.
- A validator is added at height 5.
- From now on, any call to participants(6) will return a list of 11 validators.
- This breaks the contract imposed by the Supervisor trait. For a given view, the returned list of validators has to always be the same.

Changes:
- Adds the view (from simplex) to the Block because the finalizer needs access to it.
- Uses the view instead of the block height for the registry
- Updates the registry to always find the largest view that is <= the requested view