feat: make registry updates atomic (#31)
This change makes registry updates atomic.
Currently, the registry can be updated several times for the same view: when we add participants (deposit) and when we remove participants (withdrawal).
Since the registry is in-memory, updates take places immediately, and don't have to be committed.
This change keeps track of add and removal lists for the registry and completes a single updates.
The registry uses a mutex, so the guard is held, while the registry is updated, making it atomic.

Without this change, the registry could return different validator sets for the same index.