feat: add Header to Block (#44)
This change is part of the checkpointing initiative. Every X blocks, the validators will generate a checkpoint from the consensus state and come to consensus on the checkpoint hash.
For onboarding, we need to keep track of the checkpoint signatures. To reduce the size, we don't want to include the block payload in the signature.
Therefore, we add a Header type that will be contained in the Block.

- Instead of signing the payload directly, the header contains a digest of the payload, and then the header digest is signed (the same is done for the withdrawal requests).
- When we receive a Block from the wire, we have to additionally verify that the received payload and execution requests match the payload_hash and execution_request_hash that are included in the Header.