feat: checkpointing (#45)
This introduces checkpointing for the consensus state.
An epoch is defined as a fixed number of blocks. At the end of an epoch, validators will create a checkpoint from the consensus state, and come to consensus on the checkpoint hash.

Changes:
- Introduces a new Header type that is included in the Block. The digest of this header is now the digest of the block. This allows us to keep a chain of finalized headers (which also contains the checkpoint hash) back to the genesis, without having to keep the blocks.
- Introduces an epoch, which is defined as EPOCH_NUM_BLOCKS blocks.
After committing the penultimate block of an epoch, the consensus state is written to disk and a checkpoint is created from the state.
- The hash of this checkpoint is included in the next block (the last block of the epoch).
- The syncer was updated to send the signatures together with the Block for the last block of an epoch to the application (if the signatures exist for this block).
- In the application finalizer, when the last block of the epoch is committed, the block header, together with the signatures that finalized it, is written to disk.
- e2e tests are added to verify that checkpointing works as expected