feat: re-propose last block of epoch (#76)
Re-propose the last block of an epoch to avoid producing blocks that will be cut off by the epoch boundary.
Furthermore, this handles the edge case when the last block of an epoch is notorized and nullified, because the new epoch only starts once the last block of the epoch was finalized.